refac: RefinedChildren 함수 적용
