+++ src/components/HStack/HStack.test.tsx
@@ -2,11 +2,11 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { HStack } from './HStack';
 
-const children = [<h2 id='test'>hello</h2>, <h1>hi</h1>, 7, 'elenapark'];
+const children = [<h2 id="test">hello</h2>, <h1>hi</h1>, 7, 'elenapark'];
 
 const renderStack = () => {
   const defaultProps = {
-    className: 'hstack test',
+    className: 'hstack',
     children,
   };
   render(<HStack {...defaultProps} />);
