+++ src/components/SimpleGrid/SimpleGrid.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {SimpleGrid} from './SimpleGrid';
+import { SimpleGrid } from './SimpleGrid';
 import { render, screen } from '@testing-library/react';
 
 const children = [<div>그리드1</div>, <div>그리드2</div>];
@@ -9,6 +9,7 @@ const defaultProps = {
   spacingX: 10,
   spacingY: 10,
   children,
+  className: 'simple-grid',
 };
 
 describe('<SimpleGrid />', () => {
@@ -19,7 +20,7 @@ describe('<SimpleGrid />', () => {
 
   it('should render children', () => {
     render(<SimpleGrid {...defaultProps} />);
-    expect(screen.getAllByTestId('simple-grid-item-comp')).toHaveLength(2);
+    expect(screen.getAllByTestId('simple-grid-child-comp')).toHaveLength(2);
   });
 
   it('has grid styling', () => {
